+++ tgbotapi.extensions.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/triggers_handling/MediaGroupTriggers.kt
@@ -32,7 +32,8 @@ internal suspend inline fun <reified T : MediaGroupContent> BehaviourContext.bui
     doInSubContextWithUpdatesFilter(
         updatesFilter = if (includeFilterByChatInBehaviourSubContext) {
             { it.sourceChat() ?.id ?.chatId == mediaGroupChat.id.chatId }
-        } else null
+        } else null,
+        stopOnCompletion = false
     ) {
         scenarioReceiver(mediaGroup)
     }